+++ client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for BibleEngine Client
 
+## [1.4.6]
+
+### Features
+
+-   added optional parameter `secondaryLangs` to `syncVersions`
+
 ## [1.4.0]
 
 ### Features