+++ server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for BibleEngine Server
 
+## [1.4.6]
+
+### Features
+
+-   added option to pass list of secondary languages as query parameter to `syncVersions` endpoint
+
 ## [1.4.0]
 
 ### Features